fix: перенести добавление обработчиков событий элементам карточки в функцию создания карточки
